Enlist the FreeBSD-CURRENT users as testers of what is to become Gcc 3.1.0. These bits are taken from the FSF anoncvs repo on 1-Feb-2002 08:20 PST.
Former-commit-id: 1952e2e1c1be6f107fa3ce8b10025cfd1cd7943b

+++ contrib/gcc/config/i386/svr3dbx.h
@@ -51,7 +51,7 @@ parameter_list|,
 name|NUMBER
 parameter_list|)
 define|\
-value|sprintf ((BUF), "*.%s%d", (PREFIX), (NUMBER))
+value|sprintf ((BUF), "*.%s%ld", (PREFIX), (long)(NUMBER))
 end_define
 
 begin_comment
